+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Supports for Spring Boot Actuator
 - Integrating the Spring Data Redis
+- Enhanced data processing
 
 ### 2.1.4
