+++ openyich-data/src/test/java/com/openyich/framework/data/filter/BigDecimalFilterTest.java
@@ -1,25 +1,26 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.math.BigDecimal;
 import java.util.LinkedList;
 import java.util.List;
 
 import org.junit.Before;
 import org.junit.Test;
 
+import com.openyich.framework.data.filter.BigDecimalFilter;
 import com.openyich.framework.data.filter.Filter;
-import com.openyich.framework.data.filter.IntegerFilter;
 
-public class IntegerFilterTest {
+public class BigDecimalFilterTest {
 
-  private IntegerFilter filter;
+  private BigDecimalFilter filter;
 
-  private Integer value = 42;
+  private BigDecimal value = new BigDecimal(42L);
 
   @Before
   public void setup() {
-    filter = new IntegerFilter();
+    filter = new BigDecimalFilter();
   }
 
   @Test
@@ -31,55 +32,55 @@ public class IntegerFilterTest {
     assertThat(filter.getLessOrEqualThan()).isNull();
     assertThat(filter.getSpecified()).isNull();
     assertThat(filter.getIn()).isNull();
-    assertThat(filter.toString()).isEqualTo("IntegerFilter []");
+    assertThat(filter.toString()).isEqualTo("BigDecimalFilter []");
   }
 
   @Test
   public void testSetEquals() {
-    Filter<Integer> chain = filter.setEquals(value);
+    Filter<BigDecimal> chain = filter.setEquals(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getEquals()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessThan() {
-    Filter<Integer> chain = filter.setLessThan(value);
+    Filter<BigDecimal> chain = filter.setLessThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessOrEqualThan() {
-    Filter<Integer> chain = filter.setLessOrEqualThan(value);
+    Filter<BigDecimal> chain = filter.setLessOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterThan() {
-    Filter<Integer> chain = filter.setGreaterThan(value);
+    Filter<BigDecimal> chain = filter.setGreaterThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterOrEqualThan() {
-    Filter<Integer> chain = filter.setGreaterOrEqualThan(value);
+    Filter<BigDecimal> chain = filter.setGreaterOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetSpecified() {
-    Filter<Integer> chain = filter.setSpecified(true);
+    Filter<BigDecimal> chain = filter.setSpecified(true);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getSpecified()).isEqualTo(true);
   }
 
   @Test
   public void testSetIn() {
-    List<Integer> list = new LinkedList<>();
-    Filter<Integer> chain = filter.setIn(list);
+    List<BigDecimal> list = new LinkedList<>();
+    Filter<BigDecimal> chain = filter.setIn(list);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getIn()).isEqualTo(list);
   }
@@ -94,7 +95,7 @@ public class IntegerFilterTest {
     filter.setSpecified(true);
     filter.setIn(new LinkedList<>());
     String str = value.toString();
-    assertThat(filter.toString()).isEqualTo("IntegerFilter " + "[greaterThan=" + str
+    assertThat(filter.toString()).isEqualTo("BigDecimalFilter " + "[greaterThan=" + str
         + ", greaterOrEqualThan=" + str + ", lessThan=" + str + ", " + "lessOrEqualThan=" + str
         + ", equals=" + str + ", specified=true, in=[]]");
   }
