+++ openyich-data/src/test/java/com/openyich/framework/data/filter/BooleanFilterTest.java
@@ -1,4 +1,4 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 import static org.assertj.core.api.Assertions.assertThat;
