+++ openyich-data/src/test/java/com/openyich/framework/data/filter/FloatFilterTest.java
@@ -1,8 +1,7 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.time.ZonedDateTime;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -10,17 +9,17 @@ import org.junit.Before;
 import org.junit.Test;
 
 import com.openyich.framework.data.filter.Filter;
-import com.openyich.framework.data.filter.ZonedDateTimeFilter;
+import com.openyich.framework.data.filter.FloatFilter;
 
-public class ZonedDateTimeFilterTest {
+public class FloatFilterTest {
 
-  private ZonedDateTimeFilter filter;
+  private FloatFilter filter;
 
-  private ZonedDateTime value = ZonedDateTime.now();
+  private Float value = 42F;
 
   @Before
   public void setup() {
-    filter = new ZonedDateTimeFilter();
+    filter = new FloatFilter();
   }
 
   @Test
@@ -32,55 +31,55 @@ public class ZonedDateTimeFilterTest {
     assertThat(filter.getLessOrEqualThan()).isNull();
     assertThat(filter.getSpecified()).isNull();
     assertThat(filter.getIn()).isNull();
-    assertThat(filter.toString()).isEqualTo("ZonedDateTimeFilter []");
+    assertThat(filter.toString()).isEqualTo("FloatFilter []");
   }
 
   @Test
   public void testSetEquals() {
-    Filter<ZonedDateTime> chain = filter.setEquals(value);
+    Filter<Float> chain = filter.setEquals(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getEquals()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessThan() {
-    Filter<ZonedDateTime> chain = filter.setLessThan(value);
+    Filter<Float> chain = filter.setLessThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessOrEqualThan() {
-    Filter<ZonedDateTime> chain = filter.setLessOrEqualThan(value);
+    Filter<Float> chain = filter.setLessOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterThan() {
-    Filter<ZonedDateTime> chain = filter.setGreaterThan(value);
+    Filter<Float> chain = filter.setGreaterThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterOrEqualThan() {
-    Filter<ZonedDateTime> chain = filter.setGreaterOrEqualThan(value);
+    Filter<Float> chain = filter.setGreaterOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetSpecified() {
-    Filter<ZonedDateTime> chain = filter.setSpecified(true);
+    Filter<Float> chain = filter.setSpecified(true);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getSpecified()).isEqualTo(true);
   }
 
   @Test
   public void testSetIn() {
-    List<ZonedDateTime> list = new LinkedList<>();
-    Filter<ZonedDateTime> chain = filter.setIn(list);
+    List<Float> list = new LinkedList<>();
+    Filter<Float> chain = filter.setIn(list);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getIn()).isEqualTo(list);
   }
@@ -95,8 +94,8 @@ public class ZonedDateTimeFilterTest {
     filter.setSpecified(true);
     filter.setIn(new LinkedList<>());
     String str = value.toString();
-    assertThat(filter.toString()).isEqualTo("ZonedDateTimeFilter " + "[greaterThan=" + str
-        + ", greaterOrEqualThan=" + str + ", lessThan=" + str + ", " + "lessOrEqualThan=" + str
-        + ", equals=" + str + ", specified=true, in=[]]");
+    assertThat(filter.toString()).isEqualTo(
+        "FloatFilter " + "[greaterThan=" + str + ", greaterOrEqualThan=" + str + ", lessThan=" + str
+            + ", " + "lessOrEqualThan=" + str + ", equals=" + str + ", specified=true, in=[]]");
   }
 }
