+++ openyich-data/src/test/java/com/openyich/framework/data/filter/LongFilterTest.java
@@ -1,26 +1,25 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.math.BigDecimal;
 import java.util.LinkedList;
 import java.util.List;
 
 import org.junit.Before;
 import org.junit.Test;
 
-import com.openyich.framework.data.filter.BigDecimalFilter;
 import com.openyich.framework.data.filter.Filter;
+import com.openyich.framework.data.filter.LongFilter;
 
-public class BigDecimalFilterTest {
+public class LongFilterTest {
 
-  private BigDecimalFilter filter;
+  private LongFilter filter;
 
-  private BigDecimal value = new BigDecimal(42L);
+  private Long value = 42L;
 
   @Before
   public void setup() {
-    filter = new BigDecimalFilter();
+    filter = new LongFilter();
   }
 
   @Test
@@ -32,55 +31,55 @@ public class BigDecimalFilterTest {
     assertThat(filter.getLessOrEqualThan()).isNull();
     assertThat(filter.getSpecified()).isNull();
     assertThat(filter.getIn()).isNull();
-    assertThat(filter.toString()).isEqualTo("BigDecimalFilter []");
+    assertThat(filter.toString()).isEqualTo("LongFilter []");
   }
 
   @Test
   public void testSetEquals() {
-    Filter<BigDecimal> chain = filter.setEquals(value);
+    Filter<Long> chain = filter.setEquals(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getEquals()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessThan() {
-    Filter<BigDecimal> chain = filter.setLessThan(value);
+    Filter<Long> chain = filter.setLessThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessOrEqualThan() {
-    Filter<BigDecimal> chain = filter.setLessOrEqualThan(value);
+    Filter<Long> chain = filter.setLessOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterThan() {
-    Filter<BigDecimal> chain = filter.setGreaterThan(value);
+    Filter<Long> chain = filter.setGreaterThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterOrEqualThan() {
-    Filter<BigDecimal> chain = filter.setGreaterOrEqualThan(value);
+    Filter<Long> chain = filter.setGreaterOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetSpecified() {
-    Filter<BigDecimal> chain = filter.setSpecified(true);
+    Filter<Long> chain = filter.setSpecified(true);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getSpecified()).isEqualTo(true);
   }
 
   @Test
   public void testSetIn() {
-    List<BigDecimal> list = new LinkedList<>();
-    Filter<BigDecimal> chain = filter.setIn(list);
+    List<Long> list = new LinkedList<>();
+    Filter<Long> chain = filter.setIn(list);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getIn()).isEqualTo(list);
   }
@@ -95,8 +94,8 @@ public class BigDecimalFilterTest {
     filter.setSpecified(true);
     filter.setIn(new LinkedList<>());
     String str = value.toString();
-    assertThat(filter.toString()).isEqualTo("BigDecimalFilter " + "[greaterThan=" + str
-        + ", greaterOrEqualThan=" + str + ", lessThan=" + str + ", " + "lessOrEqualThan=" + str
-        + ", equals=" + str + ", specified=true, in=[]]");
+    assertThat(filter.toString()).isEqualTo(
+        "LongFilter " + "[greaterThan=" + str + ", greaterOrEqualThan=" + str + ", lessThan=" + str
+            + ", " + "lessOrEqualThan=" + str + ", equals=" + str + ", specified=true, in=[]]");
   }
 }
