+++ openyich-data/src/test/java/com/openyich/framework/data/filter/ZonedDateTimeFilterTest.java
@@ -1,8 +1,8 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.time.Instant;
+import java.time.ZonedDateTime;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -10,17 +10,17 @@ import org.junit.Before;
 import org.junit.Test;
 
 import com.openyich.framework.data.filter.Filter;
-import com.openyich.framework.data.filter.InstantFilter;
+import com.openyich.framework.data.filter.ZonedDateTimeFilter;
 
-public class InstantFilterTest {
+public class ZonedDateTimeFilterTest {
 
-  private InstantFilter filter;
+  private ZonedDateTimeFilter filter;
 
-  private Instant value = Instant.now();
+  private ZonedDateTime value = ZonedDateTime.now();
 
   @Before
   public void setup() {
-    filter = new InstantFilter();
+    filter = new ZonedDateTimeFilter();
   }
 
   @Test
@@ -32,55 +32,55 @@ public class InstantFilterTest {
     assertThat(filter.getLessOrEqualThan()).isNull();
     assertThat(filter.getSpecified()).isNull();
     assertThat(filter.getIn()).isNull();
-    assertThat(filter.toString()).isEqualTo("InstantFilter []");
+    assertThat(filter.toString()).isEqualTo("ZonedDateTimeFilter []");
   }
 
   @Test
   public void testSetEquals() {
-    Filter<Instant> chain = filter.setEquals(value);
+    Filter<ZonedDateTime> chain = filter.setEquals(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getEquals()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessThan() {
-    Filter<Instant> chain = filter.setLessThan(value);
+    Filter<ZonedDateTime> chain = filter.setLessThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetLessOrEqualThan() {
-    Filter<Instant> chain = filter.setLessOrEqualThan(value);
+    Filter<ZonedDateTime> chain = filter.setLessOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getLessOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterThan() {
-    Filter<Instant> chain = filter.setGreaterThan(value);
+    Filter<ZonedDateTime> chain = filter.setGreaterThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetGreaterOrEqualThan() {
-    Filter<Instant> chain = filter.setGreaterOrEqualThan(value);
+    Filter<ZonedDateTime> chain = filter.setGreaterOrEqualThan(value);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getGreaterOrEqualThan()).isEqualTo(value);
   }
 
   @Test
   public void testSetSpecified() {
-    Filter<Instant> chain = filter.setSpecified(true);
+    Filter<ZonedDateTime> chain = filter.setSpecified(true);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getSpecified()).isEqualTo(true);
   }
 
   @Test
   public void testSetIn() {
-    List<Instant> list = new LinkedList<>();
-    Filter<Instant> chain = filter.setIn(list);
+    List<ZonedDateTime> list = new LinkedList<>();
+    Filter<ZonedDateTime> chain = filter.setIn(list);
     assertThat(chain).isEqualTo(filter);
     assertThat(filter.getIn()).isEqualTo(list);
   }
@@ -95,7 +95,7 @@ public class InstantFilterTest {
     filter.setSpecified(true);
     filter.setIn(new LinkedList<>());
     String str = value.toString();
-    assertThat(filter.toString()).isEqualTo("InstantFilter " + "[greaterThan=" + str
+    assertThat(filter.toString()).isEqualTo("ZonedDateTimeFilter " + "[greaterThan=" + str
         + ", greaterOrEqualThan=" + str + ", lessThan=" + str + ", " + "lessOrEqualThan=" + str
         + ", equals=" + str + ", specified=true, in=[]]");
   }
