+++ openyich-data/src/test/java/com/openyich/framework/data/mapper/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.openyich.framework.data.jpa.query;
+package com.openyich.framework.data.mapper;
 
 import javax.persistence.Column;
 import javax.persistence.GeneratedValue;
