+++ openyich-data/src/test/java/com/openyich/framework/data/mapper/ChildEntity_.java
@@ -1,4 +1,4 @@
-package com.openyich.framework.data.jpa.query;
+package com.openyich.framework.data.mapper;
 
 import javax.annotation.Generated;
 import javax.persistence.metamodel.SingularAttribute;
