+++ openyich-data/src/test/java/com/openyich/framework/data/mapper/ParentEntity.java
@@ -1,4 +1,4 @@
-package com.openyich.framework.data.jpa.query;
+package com.openyich.framework.data.mapper;
 
 import java.util.HashSet;
 import java.util.Set;
